Fix stack overflow, make sample more robust in the presense of errors
FireChannelReadComplete can call DoBeginRead recursively. However, when it happens after an exception, we end up reading 0 bytes recursively ad infinitum.
The code is brittle, but I did find a reliable repro for that SO and made sure that it's resolved after my fix. There could be other manifestations of this problem lurking elsewhere, but I could not reproduce them.