Add a filter text box to the device list
feat: Add filtering ability to device list

feat: Allow regex filters

fix: Up the max number of devices limit

There are IoT hubs that contain well more than a thousand devices.  I'm
not sure why this arbitrary limit was put in here in the first place,
but I'm assuming it was for performance.

However, to me this is a choice between the tool maybe being slow, and a
user not being able to use the tool at all to configure their device.
The choice between the two seems easy.

chore: Add region around filtering section

fix: Check for null connection strings

A new Provisioning feature allows the connection string to be null.
This wasn't being handled by the filtering previously.