Use Version instead of VersionPrefix in csproj files
VersionPrefix is only supposed to hold MAJOR.MINOR.PATCH components of a semver-compliant version. So if we want to publish a version that has a pre-release label, our current infrastructure doesn't support it. We either need to (1) parse the version to isolate the pre-release label and add it to VersionSuffix, or (2) just use Version, which is the same as VersionPrefix-VersionSuffix. This change updates our uses of VersionPrefix to Version.