feat: Limit the number of devices requested
For very large IoT Hubs, requesting every device takes an extraordinary
amount of time, and makes the tool unusable.  It takes this long because
it is making one web request per device, which is obviously horribly
inefficient.

With this commit, I'm making it possible to request a hard-coded subset
of the devices, that can be fetch in a managable amount of time.