The fix with the timeout propagates down to the HttpTransportHandler, which ensures, that there is no timeout passed in (OperationsTimeout is used there). Before the fix the timeout would never be reaching the HttpTransportHandler, but always be wrapped in a cancellation token. Which worked fine for Http but not Amqp.
Since the HttpTransportHandler relied on the handlers before it to wrap the timeout, we should be doing this inside the HttpTransportHandler and create a local cts if the passed in cts was not specified.

Since there is no public API on the DeviceClient to call the ReceiveAsync using both timeout and cancellation token, this should never happen and this change shoudl fix this behavior for the HttpTransportHandler.