- Separate open / receive methods for timeout or CTS usage - Adding OpenAsync with a timeout
The implementation of amqp does not support cancellation tokens. That's why the interfaces on the device client today, need to support a timeout.  Supporting a cancellation token in the amqp library, is more than a day project. It still uses the Begin/End async pattern, but it could be done. That would be the desired final state, where we can use the cancellation token.

Adding additional E2E testing to validate the cancellation token and timeout calls.