PR build refactor and TlsVersion instance change (#1211)
Refactored the build.ps1 script so all parameters are opt-in
Updated the cmd file that calls it to pass the right parameters
When running tests, instead of passing in a path to run, the function finds all *tests.csproj files, and starts runs with the specified filter to all projects.
This required renaming of a few test projects, and updates to the InternalsVisibleTo attributes in product assemblies.
Split up test categories, such as IotHubE2E to "IotHub" and "E2E". That way all E2E tests can be run with a single tag.
Added more test categories for tests to exclude during PR validation, such as LongRunning and Flaky
In order to get TlsVersion tests to run, I had to:

Change their test trait to match MSTest
Make the class one that can be instantiated, so it can be tested in parallel
Update calling classes to use the static Instance property
Update documentation to indicate the Instance property

Impacts Shared and most other packages.