fix(iot-dev): Fix issue where SemaphoreFullException was thrown unexpectedly
By putting the semaphore.wait() call inside a try block and the release within the finally block, the assumption is that the semaphore should always be released.

The issue here was that there are cases where the finally block is reached without ever successfully grabbing the semaphore in the first place. For instance, if the task calling into setMethodHandlerAsync is cancelled while waiting to grab the semaphore, the finally block will always throw SemaphoreFullException upon release because it never owned the semaphore