fix(iot-dev): Fix null reference exception case within RetryDelegatingHandler
If Dispose is called while this semaphore is in use, the other thread using the semaphore will hit a null reference exception when it tries to release the now null semaphore. It would be more appropriate to just hit an ObjectDisposedException instead.

In general, though, Dispose should only be called when a client is no longer open. This particular customer issue stemmed from the user disposing a client without closing it first, which opened the possibility that a connection loss event would trigger while the object is being disposed.