Fix(DeviceClient): Fix issue with hanging thread after dispose. (#2201)
Previously, in the scenario in which The `ConnectionStatusHandler` catches a disconnection and disposes the `DeviceClient` and at the same time an asynchronous SendEventAsync is waiting for the connection to be re-established, the SendEventAsync method would never return and the thread will hang for ever. 
Now the SendEventAsync method call will throw an exception indicating that the object has been disposed.