Update readme.md
Updated Readme to keep only major version listed. Customer should be able to see patches on the LTS branch and select the patch level they needed.
Changed the maintenance end date to "current" so in case the last good LTS is the current best version that will continue to receive patches.
Removed LTS End Date that is not actually enforceable from the customer perspective.