DigitalTwins subclient rework (#2534)
Removes the dependency on Microsoft.Rest.Runtime and reshapes the digital twin client's API surface to no longer use the Microsoft.Rest.Runtime classes that we used previously

This also removes all auto-generation scripts and other resources used in auto-generating this client in the past. This is being done since we likely won't autogenerate just this client again.

This also removes the folder of classes we had dedicated to implementing Microsoft.Rest.Runtime authentication primitives within the context of our service. Now we'll be using the same authentication classes as the rest of the service client subclients. These authentication methods include the track 2 friendly AzureSasCredential and TokenCredential, and
include using a simple connection string.