Move more classes out of common, delete unused common classes
FxCop.cs and Resources.Designer.cs appear to be unused. StreamDisposalResponsibility is used in both IoT Hub device and IoT Hub service clients, and TaskCompletionSource is only used in IoT Hub device client's MQTT layer.

It should be noted that TaskCompletionSource will be deleted when we move to MQTTNet and StreamDisposalResponsibility will disappear when we make the Message classes not implement IDisposable