Service client AMQP layer redo (#2614)
The v1 AMQP layer implementation was confusing and poorly abstracted from the service clients that used it. Service clients would have access to AMQP links for example. With this re-implementation, service clients will only have access to the AMQP connection and that connection will have the needed abstractions for sessions, sender links, and receiver links.

This rework also makes it so the AMQP layer is no longer disposable which lends itself to the open/close/re-open pattern that we need here.