Adopt JsonPatchDocument to replace our locally implemented UpdateOperationsUtility (#2648)
They are functionally the same, but JsonPatchDocument is managed by the Azure.Core library that we already have a dependency on

The history here is that we wrote UpdateOperationsUtility for our SDK before JsonPatchDocument existed in Azure.Core. Now that it exists in Azure.Core, there is no reason to define this class locally.