Allow users to pass in HttpClient to use for file upload operations (#3293)
Since we don't target any version of .NET core in our device SDK, there is no way for us to create an HttpClient that sets the connection lease timeout for the user by default. Instead, we will allow users to pass in an HttpClient that we will use. This allows .NET core users to construct an HTTP client with their desired connection lease timeout, and give it to the SDK to use.

For .NET standard, net472, and .NET 5+ cases, the SDK can set the connection lease timeout for the user by default.

On the service SDK side, .NET standard, net472, .NET 5+ and .NET core 2.1+ cases will create an HttpClient with the connection lease timeout set by default.